+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 Features
 
 - Default zerolog setup functions.
+- A ``NewApp()`` function to create an REST API app with logging/monitoring.
+- Add Prometheus integration to ``/metrics`` path.